+++ apps/backend/src/__tests__/routes/resource-create-validation.test.ts
@@ -22,7 +22,7 @@ import type { EventBus as EventBusType } from '@semiont/core';
 import { ResourceOperations } from '@semiont/make-meaning';
 import { registerCreateResource } from '../../routes/resources/routes/create';
 import type { ResourcesRouterType } from '../../routes/resources/shared';
-import type { ArchivistAddressConfig } from '@semiont/content';
+import type { ArchivistAddressConfig } from '@semiont/core/node';
 import { putContent } from '../../lib/archivist';
 
 vi.mock('@semiont/make-meaning', () => ({
